Module linera_alloy_rpc_types_trace::parity

Types for trace module.

See https://openethereum.github.io/JSONRPC-trace-module

Structs

AccountDiff

The diff of an account after a transaction

CallAction

Represents a certain CallType of a call or message transaction.

CallOutput

Outcome of a CALL.

ChangedType

A changed value

CreateAction

Represents a create action, either a CREATE operation or a CREATE transaction.

CreateOutput

Outcome of a CREATE.

LocalizedTransactionTrace

A wrapper for TransactionTrace that includes additional information about the transaction.

MemoryDelta

A diff of some chunk of memory.

RewardAction

Recorded reward of a block.

SelfdestructAction

Represents a selfdestruct action fka suicide.

StateDiff

New-type for list of account diffs

StorageDelta

A diff of some storage value.

TraceResults

The Outcome of a traced transaction with optional settings

TraceResultsWithTransactionHash

A FullTrace with an additional transaction hash

TransactionTrace

A parity style trace of a transaction.

VmExecutedOperation

A record of an executed VM operation.

VmInstruction

A record of a single VM instruction, opcode level.

VmTrace

A record of a full VM trace for a CALL/CREATE.

Enums

Action

Represents the various types of actions recorded during tracing

ActionType

An external action type.

CallType

Call type.

Delta

Represents how a value changed.

RewardType

What kind of reward.

TraceOutput

Represents the output of a trace.

TraceType

Different Trace diagnostic targets.